refactor: centralize isPlainObject, isRecord, isErrno, isLoopbackHost utilities (#12926)
