+++ src/browser/cdp.helpers.ts
@@ -1,6 +1,9 @@
 import WebSocket from "ws";
+import { isLoopbackHost } from "../gateway/net.js";
 import { rawDataToString } from "../infra/ws.js";
 import { getChromeExtensionRelayAuthHeaders } from "./extension-relay.js";
+
+export { isLoopbackHost };
 
 type CdpResponse = {
   id: number;
@@ -14,19 +17,6 @@ type Pending = {
 };
 
 export type CdpSendFn = (method: string, params?: Record<string, unknown>) => Promise<unknown>;
-
-export function isLoopbackHost(host: string) {
-  const h = host.trim().toLowerCase();
-  return (
-    h === "localhost" ||
-    h === "127.0.0.1" ||
-    h === "0.0.0.0" ||
-    h === "[::1]" ||
-    h === "::1" ||
-    h === "[::]" ||
-    h === "::"
-  );
-}
 
 export function getHeadersWithAuth(url: string, headers: Record<string, string> = {}) {
   const relayHeaders = getChromeExtensionRelayAuthHeaders(url);
