+++ src/channels/plugins/status-issues/shared.ts
@@ -1,9 +1,8 @@
+import { isRecord } from "../../../utils.js";
+export { isRecord };
+
 export function asString(value: unknown): string | undefined {
   return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;
-}
-
-export function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
 }
 
 export function formatMatchMetadata(params: {
