+++ src/config/includes.ts
@@ -13,6 +13,7 @@
 import JSON5 from "json5";
 import fs from "node:fs";
 import path from "node:path";
+import { isPlainObject } from "../utils.js";
 
 export const INCLUDE_KEY = "$include";
 export const MAX_INCLUDE_DEPTH = 10;
@@ -51,15 +52,6 @@ export class CircularIncludeError extends ConfigIncludeError {
 // ============================================================================
 // Utilities
 // ============================================================================
-
-function isPlainObject(value: unknown): value is Record<string, unknown> {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    !Array.isArray(value) &&
-    Object.prototype.toString.call(value) === "[object Object]"
-  );
-}
 
 /** Deep merge: arrays concatenate, objects merge recursively, primitives: source wins */
 export function deepMerge(target: unknown, source: unknown): unknown {
