+++ src/infra/errors.ts
@@ -12,6 +12,20 @@ export function extractErrorCode(err: unknown): string | undefined {
   return undefined;
 }
 
+/**
+ * Type guard for NodeJS.ErrnoException (any error with a `code` property).
+ */
+export function isErrno(err: unknown): err is NodeJS.ErrnoException {
+  return Boolean(err && typeof err === "object" && "code" in err);
+}
+
+/**
+ * Check if an error has a specific errno code.
+ */
+export function hasErrnoCode(err: unknown, code: string): boolean {
+  return isErrno(err) && err.code === code;
+}
+
 export function formatErrorMessage(err: unknown): string {
   if (err instanceof Error) {
     return err.message || err.name || "Error";
